Docker-A/requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
scikit-learn==1.6.1
Pillow==11.1.0
matplotlib==3.9.4
seaborn==0.13.2